how it all works
so there i created a channel and func where i added some variable 5 times in it. to call this func i used goroutine, so this func is running on like another level, like on another computer. if i add some functions that will compilate very slow and run them with "go" before them, they will compilate on the same time, not one after another, so it will take much less time to get the job done. but the problem is that main stop working before goroutines get the result, so we need to get our programm some time. that how i can understand it works